Revert "Fix registration"
