+++ files/buyers.csv
@@ -1,16 +1,16 @@
-adip,hacktiv8adip,adip@mail.com,jl.solo,buyer
-mamor,hacktiv8mamor,mamor@mail.com,jl.kediri,owner
-ilham,hacktiv8ilham,ilham@mail.com,jl.lombok,buyer
-irfan,hacktiv8irfan,irfan@mail.com,jl.cianjur,buyer
-bima,hacktiv8bima,bima@mail.com,jl.tulungagung,buyer
-isro,hacktiv8isro,isro@mail.com,jl.bandung,owner
-eltim,hacktiv8eltim,eltim@mail.com,jl.jakarta,buyer
-felix,hacktiv8felix,felix@mail.com,jl.medan,buyer
-ernest,hacktiv8ernest,ernest@mail.com,jl.jakarta,buyer
-ekka,hacktiv8ekka,ekka@mail.com,jl.jakarta,buyer
-ari,hacktiv8ari,ari@mail.com,jl.semarang,buyer
-sigesit,hacktiv8sigesit,sigesit@oldi.com,jl.lombok,buyer
-lalumaria,hacktiv8lalumaria,lalu@maria.com,jl.jakarta,buyer
-thereoldi,hacktiv8thereoldi,there@oldi.com,jl.jakarta,buyer
-celynalfitra,hacktiv8celynalfitra,celyn@alfitra.com,jl.jakarta,buyer
-mariacelyn,hacktiv8mariacelyn,maria@celyn.com,jl.jakarta,buyer
+adip,hacktiv8adip,adip@mail.com,jl.solo
+mamor,hacktiv8mamor,mamor@mail.com,jl.kediri
+ilham,hacktiv8ilham,ilham@mail.com,jl.lombok
+irfan,hacktiv8irfan,irfan@mail.com,jl.cianjur
+bima,hacktiv8bima,bima@mail.com,jl.tulungagung
+isro,hacktiv8isro,isro@mail.com,jl.bandung
+eltim,hacktiv8eltim,eltim@mail.com,jl.jakarta
+felix,hacktiv8felix,felix@mail.com,jl.medan
+ernest,hacktiv8ernest,ernest@mail.com,jl.jakarta
+ekka,hacktiv8ekka,ekka@mail.com,jl.jakarta
+ari,hacktiv8ari,ari@mail.com,jl.semarang
+sigesit,hacktiv8sigesit,sigesit@oldi.com,jl.lombok
+lalumaria,hacktiv8lalumaria,lalu@maria.com,jl.jakarta
+thereoldi,hacktiv8thereoldi,there@oldi.com,jl.jakarta
+celynalfitra,hacktiv8celynalfitra,celyn@alfitra.com,jl.jakarta
+mariacelyn,hacktiv8mariacelyn,maria@celyn.com,jl.jakarta
