+++ files/data.csv
@@ -1,85 +1,85 @@
-AyamLodho,TulungAgung,20000
-EntingKacang,TulungAgung,2000
-JenangGrendol,TulungAgung,7000
-JenangSabun,TulungAgung,5000
-KepitingSambal,TulungAgung
-KrupukRambak,TulungAgung
-KueSagon,TulungAgung,5000
-Sompil,TulungAgung,10000
-Kopijoss,Jogja,10000
-BakmiJogja,Jogja,15000
-Gudeg,Jogja,15000
-OsengMercon,Jogja,12000
-SegoKucing,Jogja,2000
-AyamKremes,Jogja,18000
-JadahTempe,Jogja,5000
-SateKlatak,Jogja,20000
-BlunderCokro,Madiun,5000
-KripikTempe,Madiun,10000
-KueManco,Madiun,15000
-KueSemprong,Madiun,12000
-MaduMongso,Madiun,8000
-MimosaLemon,Madiun,15000
-NasiPecel,Madiun,10000
-PentolTurah,Madiun,5000
-SambelPecel,Madiun,30000
-DawetJabun,Ponorogo,5000
-GethukGolan,Ponorogo,8000
-PecelTumpang,Ponorogo,9000
-RujakPetis,Ponorogo,7000
-Satekopok,Ponorogo,27000
-SatePonorogo,Ponorogo,15000
-SegoTahu,Ponorogo,9000
-SerabiPonorogo,Ponorogo,9000
-SegoTiwol,Ponorogo,12000
-BecekMentok,Tuban,19000
-BelutPedas,Tuban,15000
-CumiHitam,Tuban,19000
-Dumbek,Tuban,10000
-BuburSum,Tuban,12000
-GarangAsem,Tuban,13000
-Gembol,Tuban,20000
-KareRajungan,Tuban,15000
-Ampo,Tuban,19000
-OporAyam,Tuban,12000
-Serabeh,Tuban,5000
-SateBebek,Tuban,15000
-Asida,Ambon,20000
-BuburSagu,Ambon,15000
-Gogos,Ambon,18000
-IkanKuahPala,Ambon,40000
-KohuKohu,Ambon,12000
-KopiSibuSibu,Ambon,20000
-NasiKelapa,Ambon,8000
-NasiLapola,Ambon,8000
-Papeda,Ambon,10000
-RotiKeringKenari,Ambon,20000
-RujakNatsepa,Ambon,17000
-SaguWokuKomoKomo,Ambon,20000
-SambalColoColo,Ambon,30000
-TalamSaguBakar,Ambon,22000
-KerakTelor,Jakarta,20000
-Ketoprak,Jakarta,15000
-KueSagu,Jakarta,20000
-LaksaBetawi,Jakarta,22000
-LontongSayur,Jakarta,20000
-NasiUduk,Jakarta,23000
-NasiUlam,Jakarta,20000
-SemurJengkol,Jakarta,10000
-SotoBetawi,Jakarta,20000
-SotoTangkar,Jakarta,20000
-AsinanBogor,Sunda,18000
-Awug,Sunda,8000
-AyamBakaka,Sunda,30000
-Batagor,Sunda,10000
-Cilok,Sunda,7000
-DodolGarut,Sunda,8000
-EmpalGentong,Sunda,20000
-MieKocok,Sunda,13000
-NasiTimbel,Sunda,20000
-Peuyeum,Sunda,10000
-SateMaranggi,Sunda,50000
-Seblak,Sunda,20000
-SotoKuningBogor,Sunda,30000
-TahuSumedang,Sunda,20000
-Wajit,Sunda,6000
+AyamLodho,TulungAgung,20000,pedas
+EntingKacang,TulungAgung,2000,manis
+JenangGrendol,TulungAgung,7000,manis
+JenangSabun,TulungAgung,5000,manis
+KepitingSambal,TulungAgung,17000,pedas
+KrupukRambak,TulungAgung,10000,gurih
+KueSagon,TulungAgung,5000,manis
+Sompil,TulungAgung,10000,gurih
+Kopijoss,Jogja,10000,pahit
+BakmiJogja,Jogja,15000,lezat
+Gudeg,Jogja,15000,manis
+OsengMercon,Jogja,12000,pedas
+SegoKucing,Jogja,2000,pedas
+AyamKremes,Jogja,18000,pedas
+JadahTempe,Jogja,5000,pedas
+SateKlatak,Jogja,20000,pedas
+BlunderCokro,Madiun,5000,pedas
+KripikTempe,Madiun,10000,gurih
+KueManco,Madiun,15000,manis
+KueSemprong,Madiun,12000,manis
+MaduMongso,Madiun,8000,manis
+MimosaLemon,Madiun,15000,manis
+NasiPecel,Madiun,10000,pedas
+PentolTurah,Madiun,5000,pedas
+SambelPecel,Madiun,30000,pedas
+DawetJabun,Ponorogo,5000,manis
+GethukGolan,Ponorogo,8000,manis
+PecelTumpang,Ponorogo,9000,pedas
+RujakPetis,Ponorogo,7000,pedas
+Satekopok,Ponorogo,27000,gurih
+SatePonorogo,Ponorogo,15000,gurih
+SegoTahu,Ponorogo,9000,gurih
+SerabiPonorogo,Ponorogo,9000,gurih
+SegoTiwol,Ponorogo,12000,gurih
+BecekMentok,Tuban,19000,lezat
+BelutPedas,Tuban,15000,pedas
+CumiHitam,Tuban,19000,pedas
+Dumbek,Tuban,10000,gurih
+BuburSum,Tuban,12000,lezat
+GarangAsem,Tuban,13000,asam
+Gembol,Tuban,20000,manis
+KareRajungan,Tuban,15000,manis
+Ampo,Tuban,19000,manis
+OporAyam,Tuban,12000,lezat
+Serabeh,Tuban,5000,gurih
+SateBebek,Tuban,15000,gurih
+Asida,Ambon,20000,manis
+BuburSagu,Ambon,15000,gurih
+Gogos,Ambon,18000,lezat
+IkanKuahPala,Ambon,40000,gurih
+KohuKohu,Ambon,12000,pedas
+KopiSibuSibu,Ambon,20000,manis
+NasiKelapa,Ambon,8000,gurih
+NasiLapola,Ambon,8000,lezat
+Papeda,Ambon,10000,manis
+RotiKeringKenari,Ambon,20000,lezat
+RujakNatsepa,Ambon,17000,pedas
+SaguWokuKomoKomo,Ambon,20000,manis
+SambalColoColo,Ambon,30000,pedas
+TalamSaguBakar,Ambon,22000,lezat
+KerakTelor,Jakarta,20000,gurih
+Ketoprak,Jakarta,15000,pedas
+KueSagu,Jakarta,20000,manis
+LaksaBetawi,Jakarta,22000,lezat
+LontongSayur,Jakarta,20000,pedas
+NasiUduk,Jakarta,23000,pedas
+NasiUlam,Jakarta,20000,gurih
+SemurJengkol,Jakarta,10000,pedas
+SotoBetawi,Jakarta,20000,lezat
+SotoTangkar,Jakarta,20000,gurih
+AsinanBogor,Sunda,18000,lezat
+Awug,Sunda,8000,manis
+AyamBakaka,Sunda,30000,pedas
+Batagor,Sunda,10000,gurih
+Cilok,Sunda,7000,pedas
+DodolGarut,Sunda,8000,manis
+EmpalGentong,Sunda,20000,lezat
+MieKocok,Sunda,13000,lezat
+NasiTimbel,Sunda,20000,pedas
+Peuyeum,Sunda,10000,manis
+SateMaranggi,Sunda,50000,gurih
+Seblak,Sunda,20000,pedas
+SotoKuningBogor,Sunda,30000,lezat
+TahuSumedang,Sunda,20000,gurih
+Wajit,Sunda,6000,manis
