--- models/foodtag.js
@@ -1,11 +1,0 @@
-'use strict';
-module.exports = (sequelize, DataTypes) => {
-  const Foodtag = sequelize.define('Foodtag', {
-    tagId: DataTypes.INTEGER,
-    foodId: DataTypes.INTEGER
-  }, {});
-  Foodtag.associate = function(models) {
-    
-  };
-  return Foodtag;
-};
