+++ views/register/register.ejs
@@ -6,14 +6,11 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>@dang Register</title>
-  <%- include('../partials/cssNavBar.ejs') %>
 </head>
 
 <body>
-  <%- include('../partials/notif.ejs') %>
-  <%- include('../partials/navBar.ejs') %>
   @dang Register
-  <h1> Register </h1>
+  <h1>register </h1>
   <form action="/register" method="POST">
     <br>
     username: <input type="text" name="username">
@@ -24,7 +21,8 @@
     <br>
     address: <input type="text" name="address">
     <br>
-    <button type="submit">Submit</button>
+    deposit: <input type="text" name="deposit">
+    <button type="submit">Masukkan</button>
   </form>
 
 </body>
